Extract InteractionButtons component
